services/web/requirements.txt
click==8.1.3
cycler==0.11.0
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.34.4
gunicorn==20.1.0
importlib-metadata==4.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
pandas==1.4.3
patsy==0.5.2
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
Werkzeug==2.1.2
xlrd==2.0.1
zipp==3.8.1
